Serverless/models/image-to-text/requirements.txt
functions-framework==3.*
google-cloud-storage==1.25.0
transformers==4.38.2
tensorflow
pillow